[editor] split API unit tests to separate files
