+++ editor/__tests__/mocks/editor-factory.js
