+++ editor/__tests__/mocks/webpack-resolver.js
