+++ editor/__tests__/setup.js
@@ -1,0 +1,9 @@
+/* globals beforeEach, jest */
+beforeEach(() => {
+  navigator.qt = {
+    postMessage: jest.fn()
+  }
+
+  delete window.location
+  window.location = { assign: jest.fn() }
+})
